Fix pylint errors
